requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.20
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
openai==1.58.1
Pillow==11.0.0
python-slugify==8.0.4
feedgen==1.0.0
aiofiles==24.1.0
